feat i20221031 able to run with hardcoded values
Adds a java program that can run when a person writes the url,
username, password, and other configuring values in RfdInit.
That satisfies the issue, so I've closed it. I expect to open another
for providing the credentials via another mechanism.

I chose java for implementation ease (over go or bash), given that
we have a jre everywhere I intend to use it.