brk chg i2023-02-17 uses json file for database
This changes the primary means of providing the database credentials
from hand typed to consuming a json file. It will, however, fall back
to requesting the values individually if the json file isn't appropriate
or accessible.

This no longer accepts a command line argument. Instead, it asks
for three values in the simplest case : the sql folder, the json file,
and the mysql executable. That way, it is consistent.